requirements.txt
pyspark==3.1.2
pytest==6.2.5
requests
black
python-dotenv
boto3
